Moar binding lib goodness.  Properties work again.  Fixed some bugs.
Also I'm trying out a new behavior of weak references where they will become null at some undefined point in the future when the object they reference becomes null.  Also, if a weak reference becomes null, and it is the key or value of a table, that key-value pair will be removed, again at some undefined point in the future.  This is mostly a trial, it might not stay around.