Closes #93, #95, and #96. Added scope action statements (scope(success) and the like), like in D. While testing them I found a veritable hornet's nest of issues with exception handling codegen and interpretation.  Probably because I wrote that code over 2 years ago and haven't looked at it since.
I'll be putting much clearer docs on the semantics of try-catch-finally (and the scope statements, of course) in the wiki.  For now, you need to know that break, continue, and return are now forbidden within finally blocks, and that "return foo(bar)" is not turned into a tailcall inside try and catch blocks.