The 'as' expression returns from the ashes in a new form! Instead of checking class inheritance though, it (for now) lets you typecast to bool, int, float, and string. This replaces the toBool(), toInt(), toFloat(), and toString() functions in the standard library (though toString is still there since it's likely to show up in a lot of code).
For now there's no operator overloading for 'as', and you can't use user types, but those are likely to show up in the
future.

This also closes #73 (duplication of toInt and string.toInt) since 'as int' and 'as float' do not do string parsing,
meaning string.toInt and string.toFloat are the only way to parse numbers out of strings now.